refactor(api): split sse.go by concern (hub / serve / rate-limiter)
sse.go grew to 563 lines mixing types, the hub event loop, HTTP serving,
and the rate limiter. Split into four files so each piece stays well
under the project's red-flag size budget and the next pass on the SSE
log tee (which needs to touch the hub for a reentrancy guard) doesn't
have to navigate around unrelated code.

Layout:
- sse.go (123)             — types, constants, SSEConfig
- sse_hub.go (249)         — SSEHub lifecycle + broadcast API
- sse_serve.go (209)       — HTTP serving + per-connection event pump
- sse_rate_limiter.go (60) — per-stream token bucket

Pure code move; behaviour and method signatures are unchanged. Build,
vet, and go test ./internal/api/... all green.
