+++ internal/api/sse_hub.go
@@ -1,0 +1,249 @@
+package api
+
+import (
+	"encoding/json"
+	"fmt"
+	"log/slog"
+	"time"
+)
+
+// SSEHub manages connected SSE clients and dispatches broadcasts to
+// the right per-stream subset. The hub owns a Run goroutine that
+// serialises register / unregister / broadcast events through unbuffered
+// channels, which keeps the per-stream client map race-free without
+// requiring callers to hold a lock.
+//
+// Lifecycle:
+//   - NewSSEHub() builds the hub. Call Run() in a goroutine.
+//   - HTTP handlers send on hub.register / hub.unregister.
+//   - Producers (packet observer, daemon stats, future log tee) call
+//     Broadcast / BroadcastPacket / BroadcastLog / BroadcastStats.
+//   - Stop() closes stopChan; Run drains and unregisters every client.
+
+// NewSSEHub creates a new SSE hub with the given configuration. Zero
+// values in cfg are replaced by defaults from SSEConfig.withDefaults.
+func NewSSEHub(cfg SSEConfig) *SSEHub {
+	cfg = cfg.withDefaults()
+	hub := &SSEHub{
+		clients:      make(map[SSEStream]map[*SSEClient]bool),
+		broadcast:    make(chan *streamMessage, cfg.BufferSize),
+		register:     make(chan *SSEClient),
+		unregister:   make(chan *SSEClient),
+		rateLimiters: make(map[SSEStream]*sseRateLimiter),
+		stopChan:     make(chan struct{}),
+		config:       cfg,
+	}
+
+	// Initialize per-stream structures
+	for _, stream := range []SSEStream{SSEStreamPackets, SSEStreamLogs, SSEStreamStats} {
+		hub.clients[stream] = make(map[*SSEClient]bool)
+		hub.rateLimiters[stream] = newSSERateLimiter(int64(cfg.MaxMsgPerSec))
+	}
+
+	return hub
+}
+
+// Run starts the hub's main event loop. Blocks until Stop is called.
+// Run on its own goroutine — typically launched once at daemon startup.
+func (h *SSEHub) Run() {
+	h.running.Store(true)
+	defer h.running.Store(false)
+
+	for {
+		select {
+		case client := <-h.register:
+			h.registerClient(client)
+
+		case client := <-h.unregister:
+			h.unregisterClient(client)
+
+		case msg := <-h.broadcast:
+			h.broadcastMessage(msg)
+
+		case <-h.stopChan:
+			h.closeAllClients()
+
+			return
+		}
+	}
+}
+
+// Stop gracefully shuts down the hub. Safe to call multiple times.
+func (h *SSEHub) Stop() {
+	h.stopOnce.Do(func() {
+		close(h.stopChan)
+	})
+}
+
+func (h *SSEHub) registerClient(client *SSEClient) {
+	logger := slog.Default()
+	h.mu.Lock()
+	defer h.mu.Unlock()
+
+	streamClients := h.clients[client.stream]
+
+	if len(streamClients) >= h.config.MaxClients {
+		logger.Warn(
+			"[SSE] Rejecting client for stream: max clients reached",
+			"stream",
+			client.stream,
+			"maxClients",
+			h.config.MaxClients,
+		)
+		client.closed.Store(true)
+		close(client.send)
+
+		return
+	}
+
+	streamClients[client] = true
+	logger.Info(
+		"[SSE] Client connected to stream",
+		"stream",
+		client.stream,
+		"clientIP",
+		client.clientIP,
+		"total",
+		len(streamClients),
+	)
+}
+
+func (h *SSEHub) unregisterClient(client *SSEClient) {
+	logger := slog.Default()
+	h.mu.Lock()
+	defer h.mu.Unlock()
+
+	streamClients := h.clients[client.stream]
+	if _, ok := streamClients[client]; ok {
+		delete(streamClients, client)
+
+		if !client.closed.Load() {
+			client.closed.Store(true)
+			close(client.send)
+		}
+
+		logger.Info("[SSE] Client disconnected from stream", "stream", client.stream, "remaining", len(streamClients))
+	}
+}
+
+func (h *SSEHub) broadcastMessage(msg *streamMessage) {
+	h.mu.RLock()
+	defer h.mu.RUnlock()
+
+	rateLimiter := h.rateLimiters[msg.stream]
+	if rateLimiter != nil && !rateLimiter.allow() {
+		return // Rate limit exceeded
+	}
+
+	streamClients := h.clients[msg.stream]
+	for client := range streamClients {
+		if client.closed.Load() {
+			continue
+		}
+
+		select {
+		case client.send <- msg.data:
+		default:
+			// Buffer full, drop message
+		}
+	}
+}
+
+func (h *SSEHub) closeAllClients() {
+	logger := slog.Default()
+	h.mu.Lock()
+	defer h.mu.Unlock()
+
+	for stream, clients := range h.clients {
+		for client := range clients {
+			if !client.closed.Load() {
+				client.closed.Store(true)
+				close(client.send)
+			}
+		}
+
+		h.clients[stream] = make(map[*SSEClient]bool)
+	}
+
+	logger.Info("[SSE] All clients disconnected")
+}
+
+// Broadcast sends a message to all clients of a stream. Messages are
+// dropped if either the broadcast channel or the per-stream rate
+// limiter is saturated — the hub favours liveness over delivery
+// guarantees, since SSE clients auto-reconnect and can refetch state.
+func (h *SSEHub) Broadcast(stream SSEStream, data any) {
+	logger := slog.Default()
+	if !h.running.Load() {
+		return
+	}
+
+	// Increment event ID
+	eventID := h.eventID.Add(1)
+
+	// Format as SSE
+	jsonData, err := json.Marshal(data)
+	if err != nil {
+		logger.Error("[SSE] Failed to marshal message", "error", err)
+
+		return
+	}
+
+	// SSE format: "id: N\ndata: {json}\n\n"
+	sseData := fmt.Sprintf("id: %d\ndata: %s\n\n", eventID, jsonData)
+
+	select {
+	case h.broadcast <- &streamMessage{stream: stream, data: []byte(sseData)}:
+	default:
+		// Broadcast channel full
+	}
+}
+
+// BroadcastPacket sends a packet to all packet stream subscribers.
+func (h *SSEHub) BroadcastPacket(data any) {
+	h.Broadcast(SSEStreamPackets, map[string]any{
+		"type":      "packet",
+		"data":      data,
+		"timestamp": time.Now().UTC(),
+	})
+}
+
+// BroadcastLog sends a log message to all log stream subscribers.
+func (h *SSEHub) BroadcastLog(level, message string) {
+	h.Broadcast(SSEStreamLogs, map[string]any{
+		"type":      "log",
+		"level":     level,
+		"message":   message,
+		"timestamp": time.Now().UTC(),
+	})
+}
+
+// BroadcastStats sends statistics to all stats stream subscribers.
+func (h *SSEHub) BroadcastStats(data any) {
+	h.Broadcast(SSEStreamStats, map[string]any{
+		"type":      "stats",
+		"data":      data,
+		"timestamp": time.Now().UTC(),
+	})
+}
+
+// ClientCount returns the number of clients for a stream.
+func (h *SSEHub) ClientCount(stream SSEStream) int {
+	h.mu.RLock()
+	defer h.mu.RUnlock()
+
+	return len(h.clients[stream])
+}
+
+// TotalClientCount returns total connected clients across every stream.
+func (h *SSEHub) TotalClientCount() int {
+	h.mu.RLock()
+	defer h.mu.RUnlock()
+
+	total := 0
+	for _, clients := range h.clients {
+		total += len(clients)
+	}
+
+	return total
+}
